+++ src/FunFair.Test.Common/Logging/DisposableLogger.cs
@@ -1,0 +1,37 @@
+using System;
+using Microsoft.Extensions.Logging;
+
+namespace FunFair.Test.Common
+{
+    internal sealed class DisposableLogger : ILogger, IDisposable
+    {
+        private readonly ILogger _logger;
+        private readonly IDisposable _scope;
+
+        public DisposableLogger(ILogger logger)
+        {
+            this._logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            this._scope = this._logger.BeginScope(state: "Test");
+        }
+
+        public void Dispose()
+        {
+            this._scope.Dispose();
+        }
+
+        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+            this._logger.Log(logLevel, eventId, state, exception, formatter);
+        }
+
+        public bool IsEnabled(LogLevel logLevel)
+        {
+            return this._logger.IsEnabled(logLevel);
+        }
+
+        public IDisposable BeginScope<TState>(TState state)
+        {
+            return this._logger.BeginScope(state);
+        }
+    }
+}