+++ src/FunFair.Test.Common/Logging/XUnitLoggingHelpers.cs
@@ -1,0 +1,17 @@
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.Extensions.Logging;
+using Xunit.Abstractions;
+
+namespace FunFair.Test.Common.Logging
+{
+    internal static class XUnitLoggingHelpers
+    {
+        [SuppressMessage(category: "Microsoft.Reliability", checkId: "CA2000:DisposeObjectsBeforeLosingScope", Justification = "A mock of unit tests")]
+        public static ILoggingBuilder AddXUnit(this ILoggingBuilder builder, ITestOutputHelper output)
+        {
+            builder.AddProvider(new XunitLoggerProvider(output));
+
+            return builder;
+        }
+    }
+}