+++ src/FunFair.Test.Common/Startup/LoggingStartup.cs
@@ -1,4 +1,4 @@
-using FunFair.Test.Common.Helpers;
+using FunFair.Test.Common.Logging;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Xunit.Abstractions;